Implement the add functionality to dom and localstorage
