+++ src/index.html
@@ -19,8 +19,12 @@
           </header>
 
           <div class="user__input space__between">
-            <input type="text" placeholder="Add to your list..." />
-            <i class="fa-solid fa-arrow-left-long"></i>
+            <input
+              type="text"
+              id="user__input__field"
+              placeholder="Add to your list..."
+            />
+            <i class="fa-solid fa-arrow-left-long" id="user__input__btn"></i>
           </div>
 
           <ul id="items__container"></ul>
